+++ Levi9.ERP.Domain/Repository/PriceListRepository.cs
@@ -11,13 +11,14 @@ namespace Levi9.ERP.Domain.Repository
 {
     public class PriceListRepository : IPriceListRepository
     {
-        private DataBaseContext _dataBaseContext;
+        private readonly DataBaseContext _dataBaseContext;
         public PriceListRepository(DataBaseContext dataBaseContext) 
         {
             _dataBaseContext = dataBaseContext;
         }
         public async Task<PriceList> GetByIdAsync(int id)
         {
+            //return await _dataBaseContext.PriceLists.FirstOrDefaultAsync(p => p.Id == id);
             return await _dataBaseContext.PriceLists.FirstAsync(p => p.Id == id);
         }
     }
