+++ Levi9.ERP/Controllers/PriceListController.cs
@@ -5,34 +5,44 @@ using Levi9.ERP.Domain.Model;
 using Levi9.ERP.Domain.Model.DTO;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Routing;
 
 namespace Levi9.ERP.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("v1/[controller]")]
     [ApiController]
-    public class PriceListController : ControllerBase
+    public class PricelistController : ControllerBase
     {
         private readonly IPriceListService _priceListService;
         private readonly IMapper _mapper;
-        public PriceListController(IPriceListService priceListService, IMapper mapper) 
+        public PricelistController(IPriceListService priceListService, IMapper mapper) 
         {
             _priceListService = priceListService;
             _mapper = mapper;
         }
 
-        [HttpGet("id")]
+        [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id) 
         {
+            /*
+            var priceDTO = await _priceListService.GetByIdAsync(id);
+
+            if(priceDTO == null) 
+                return NotFound($"Nonexistent price list with ID: {id}"); 
+            
+            var priceListResponse = _mapper.Map<PriceListResponse>(priceDTO);
+
+            return Ok(priceListResponse);
+            */
             try
             {
-
                 var priceDTO = await _priceListService.GetByIdAsync(id);
                 var priceListResponse = _mapper.Map<PriceListResponse>(priceDTO);
                 return Ok(priceListResponse);
             }
             catch(InvalidOperationException)
             {
-                return BadRequest($"Non existent price list with ID: {id}");
+                return NotFound($"Nonexistent price list with ID: {id}");
             }
             catch(ArgumentException)
             {
